refactor: hooks 폴더 구조 변경
1. hooks 바로 내부의 구조를 'queries'와 'pages'의 폴더 구조로 변경
- 'queries' : react-query를 사용하여 서버 상태값을 호출하기 위한 커스텀
  훅
- 'pages' : 특정 페이지의 비즈니스 로직을 처리하기 위한 커스텀 훅
- 상기의 2개의 폴더 구조로 변경함으로써, 하기의 이득을 얻을 수 있을
  것이라 예상함
  1. react-query와 관련된 코드 내 변경이 발생했을 때, 사용한 코드를
     헤집어 들어가는 것이 아니라 'queries' 내부의 폴더를 수정하는 것으로
     쉽게 해당 코드를 수정할 수 있음
  2. 커스텀 훅이 산재되어 각각의 훅들의 의미를 제목 이외에는 쉽게
     판단하기 어려웠지만, 폴더를 통한 분류화를 통해 폴더/파일의
     의미를 더 파악하기 쉬워짐
