+++ FE/src/hooks/pages/useDetail.tsx
